% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_computation.R
\name{area_computation}
\alias{area_computation}
\title{Compute area statistics for region of interest on desired polygon datasets}
\usage{
area_computation(aoi = NULL, target_data = NULL, unit = "m")
}
\arguments{
\item{aoi}{area of interest - polygon object (sf)}

\item{target_data}{An sf object of which area is to be computed}

\item{unit}{A character vector defining whether the result area should be in meter ("m") or kilometer ("km") - meter being default value}
}
\value{
A dataframe with areal statistics
}
\description{
Compute area statistics for region of interest on desired polygon datasets
}
\examples{

# load wdpa polygon
pa <- sf::st_read(system.file("extdata", "test_wdpa.gpkg", package = "openRspat"))
# load test data
teow <- sf::st_read(system.file("extdata", "test_eoregions.gpkg", package = "openRspat"))
# define unit
unit <- "m"
# call function
area_computation(aoi=pa, target_data=teow, unit=unit)

}
