% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_proj.R
\name{area_proj}
\alias{area_proj}
\title{Area Projection}
\usage{
area_proj(x)
}
\arguments{
\item{x}{sf object}
}
\value{
LAEA projection string
}
\description{
Get LAEA projection strings from the provided sf object
}
\examples{
fname <- system.file("shape/nc.shp", package="sf")
sf <- sf::st_read(fname)
x <- sf[1, ]
area_proj(x)
}
