% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_accessibility.R
\name{compute_accessibility}
\alias{compute_accessibility}
\title{Compute zonal statistics for travel time from the area of interest to the nearby cities}
\usage{
compute_accessibility(aoi = NULL, acc_rast = NULL, idcol = NULL, opn = "mean")
}
\arguments{
\item{aoi}{area of interest - polygon (sf)}

\item{acc_rast}{accessibility raster (terra)}

\item{idcol}{A character vector identifying a column name which uniquely
identifies the polygons in the \code{aoi} object.}

\item{opn}{A type of zonal operation (sum/mean/min/max) - mean being default value}
}
\value{
zonal statistics
}
\description{
Returns the result in long-table format
}
\examples{

\donttest{
# load raster
r <- terra::rast(system.file("extdata", "test_acc_rast.tif", package = "openRspat"))
# load polygon
fname <- system.file("shape/nc.shp", package="sf")
sf <- sf::st_read(fname)
# provide unique id column
idcol <- sf[1, ]$CNTY_ID
# zonal operation type
opn <- "min"
# call function
compute_accessibility(aoi=sf[1,], acc_rast=r, idcol=idcol, opn=opn)
}

}
