% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_copernicus_land_cover.R
\name{compute_land_cover}
\alias{compute_land_cover}
\title{Compute land cover statistics}
\usage{
compute_land_cover(r, p, y)
}
\arguments{
\item{r}{land cover raster}

\item{p}{polygon object (sf)}

\item{y}{year}
}
\value{
land cover classes
}
\description{
Compute copernicus global land cover statistics - 23 discrete classifications
}
\examples{

# load raster
r <- terra::rast(system.file("extdata", "test_lc_rast.tif", package = "openRspat"))
# load polygon
fname <- system.file("shape/nc.shp", package="sf")
sf <- sf::st_read(fname)
# transform sf object to match projection with raster object
sf <- sf::st_transform(sf, "+proj=longlat +datum=WGS84 +no_defs")
# call function
compute_land_cover(r, sf[1,], 2015)

}
