% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clay_content.R
\name{get_clay_content}
\alias{get_clay_content}
\title{Download global clay content rasters at three different depths}
\usage{
get_clay_content(depth = NULL, file_path = NULL)
}
\arguments{
\item{depth}{depth in cm (0, 10, 30)}

\item{file_path}{path location}
}
\value{
clay content rasters
}
\description{
Download global clay content rasters at three different depths
}
\examples{

\donttest{
file_path <- ".." # path to the folder where you want to download the rasters
# Download population count rasters
get_clay_content(depth = 30, file_path=file_path)
}

}
