% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_climate_variables.R
\name{get_climate_rasters}
\alias{get_climate_rasters}
\title{Get 30s climate raster datasets - average temperature and precipitation (1970-2000)}
\usage{
get_climate_rasters(var = NULL, file_path = NULL)
}
\arguments{
\item{var}{use "tavg" for average temperature rasters and "prec" for precipitation rasters - pass as character value}

\item{file_path}{path location}
}
\value{
climatic variable rasters
}
\description{
Get 30s climate raster datasets - average temperature and precipitation (1970-2000)
}
\examples{

\donttest{
file_path <- ".." # path to the folder where you want to download the rasters
# Download average temperature rasters
get_climate_rasters(var = "tavg", file_path=file_path)
}

}
