% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_copernicus_land_cover.R
\name{get_copernicus_land_cover}
\alias{get_copernicus_land_cover}
\title{Get Copernicus Global Land Cover}
\usage{
get_copernicus_land_cover(year = NULL, grid = NULL, file_path = NULL)
}
\arguments{
\item{year}{year (2015 to 2019)}

\item{grid}{Latlon grid}

\item{file_path}{path}
}
\value{
copernicus land cover rasters
}
\description{
Download Copernicus Global Land Cover rasters with 23 discrete classifications for the desired year from 2015 to 2019 and for the desired grid
}
\examples{
f <- ".." # path to the folder where you want to download the rasters
get_copernicus_land_cover(year=2015, grid="W120N40", file_path=f)
}
