% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecoregions.R
\name{get_ecoregions}
\alias{get_ecoregions}
\title{Download terrestrial ecoregions of the world polygon}
\usage{
get_ecoregions(file_path = NULL)
}
\arguments{
\item{file_path}{path location}
}
\value{
teow gpkg
}
\description{
It is a one-time downloadable data. The function first downloads the zip file, unzip it, load the shapefile, convert the shapefile to geopackage and write the geopackage to file.
}
\examples{

\donttest{
file_path <- ".." # path to the folder where you want to download the polygon
# Download teow polygon
get_ecoregions(file_path=file_path)
}

}
