% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gadm.R
\name{get_gadm_gpkg}
\alias{get_gadm_gpkg}
\title{Get GADM country level datasets (geopackage)}
\usage{
get_gadm_gpkg(iso = NULL, file_path = NULL)
}
\arguments{
\item{iso}{ISO3 country code}

\item{file_path}{filepath}
}
\value{
gadm geopackage
}
\description{
Download country level administrative datasets in geopackage
}
\examples{
# give path to your destination folder
file_path <- ".."
# call the function
get_gadm_gpkg("NPL", f)

}
