% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gadm_shp.R
\name{get_gadm_shp}
\alias{get_gadm_shp}
\title{Get GADM country level datasets (shapefile)}
\usage{
get_gadm_shp(x, f)
}
\arguments{
\item{x}{ISO3 country code}

\item{f}{filepath}
}
\value{
gadm shapefiles
}
\description{
Download country level administrative datasets in ESRI shapefile format
}
\examples{
# give path to your destination folder
f <- "../../"
# call the function
get_gadm_shp("NPL", f)

}
