% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_mangrove_watch.R
\name{get_global_mangrove_watch}
\alias{get_global_mangrove_watch}
\title{Download global mangrove watch polygon data}
\usage{
get_global_mangrove_watch(yyyy = NULL, file_path = NULL)
}
\arguments{
\item{yyyy}{year (1996, 2007, 2008, 2009, 2010, 2015, 2016)}

\item{file_path}{path location}
}
\value{
gmw gpkg
}
\description{
The function first downloads the zip file, unzip it, load the shapefile...
...convert the shapefile to geopackage and write the geopackage to file.
}
\examples{

\donttest{
file_path <- ".." # path to the folder where you want to download the polygon
# Download teow polygon
get_global_mangrove_watch(yyyy = 2007, file_path=file_path)
}

}
