% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_net_forest_carbon_flux.R
\name{get_net_forest_carbon_flux}
\alias{get_net_forest_carbon_flux}
\title{Get Net Forest Carbon Flux rasters (carbon balance)}
\usage{
get_net_forest_carbon_flux(grid = NULL, file_path = NULL)
}
\arguments{
\item{grid}{10*10 Latlon grid (e.g. "00N_040W")}

\item{file_path}{path location}
}
\value{
carbon balance rasters
}
\description{
Download Net Forest Carbon Flux rasters from Global Forest Watch (GFW) for the desired grid
Please register at GFW first from your PC in order to access the API to raster files
}
\examples{

\donttest{
file_path <- ".." # path to the folder where you want to download the rasters
# Download carbon balance rasters for the grid 00N,040W
get_net_forest_carbon_flux(grid="00N_040W", file_path=file_path)
}

}
