% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_count.R
\name{get_pop_count}
\alias{get_pop_count}
\title{Get 1km global population count rasters}
\usage{
get_pop_count(yyyy = NULL, file_path = NULL)
}
\arguments{
\item{yyyy}{year (from 2000 to 2020)}

\item{file_path}{path location}
}
\value{
population count rasters
}
\description{
Get 1km global population count rasters
}
\examples{

\donttest{
file_path <- ".." # path to the folder where you want to download the rasters
# Download population count rasters
get_pop_count(yyyy = 2020, file_path=file_path)
}

}
