% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openRspat.R
\docType{package}
\name{openRspat}
\alias{openRspat}
\title{Open source geo-spatial analysis}
\description{
Contain functions to download free & open geo-datasets, convert projection of the sf objects, and geospatial analysis of the datasets.
}
\author{
Om Prakash Bhandari \email{om.bhandari1995@gmail.com}
}
