% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_operation.R
\name{zonal_operation}
\alias{zonal_operation}
\title{Compute zonal statistics from provided raster files}
\usage{
zonal_operation(aoi = NULL, idcol = NULL, rast = NULL, opn = NULL)
}
\arguments{
\item{aoi}{An sf object with polygons representing areas of interest}

\item{idcol}{A character vector identifying a column name which uniquely
identifies the polygons in the \code{aoi} object.}

\item{rast}{A raster object as spatRaster on which to perform zonal operation}

\item{opn}{A type of zonal operation (e.g. sum, mean, min, max)}
}
\value{
A dataframe with zonal statistics
}
\description{
Compute zonal statistics from provided raster files
}
\examples{

# load raster
rast <- terra::rast(system.file("extdata", "test_worldpop2020.tif", package = "openRspat"))
# load polygon
aoi <- sf::st_read(system.file("extdata", "test_wdpa.gpkg", package = "openRspat"))
# provide unique id column
idcol <- "WDPAID"
# zonal operation type
opn <- "min"
# call function
zonal_operation(aoi, idcol, rast, opn)

}
