% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_computation.R
\name{area_computation}
\alias{area_computation}
\title{Compute area statistics for region of interest on desired polygon datasets}
\usage{
area_computation(roi, data, unit)
}
\arguments{
\item{roi}{An sf object with polygons representing region of interest}

\item{data}{An sf object of which area is to be computed}

\item{unit}{A character vector defining whether the result area should be in meter ("m") or kilometer ("km")}
}
\value{
A dataframe with areal statistics
}
\description{
Compute area statistics for region of interest on desired polygon datasets
}
\examples{

# load wdpa polygon
pa <- sf::st_read(system.file("extdata", "test_wdpa.gpkg", package = "openRspat"))
# load test data
teow <- sf::st_read(system.file("extdata", "test_ecoregions.gpkg", package = "openRspat"))
# define unit
unit <- "m"
# call function
area_computation(pa, teow, unit)

}
