% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_computation.R
\name{area_proj}
\alias{area_proj}
\title{Area Projection}
\usage{
area_proj(aoi = NULL)
}
\arguments{
\item{aoi}{area of interest - polygon object (sf)}
}
\value{
LAEA projection string
}
\description{
Get LAEA projection strings from the provided sf object
}
\examples{
fname <- system.file("shape/nc.shp", package="sf")
sf <- sf::st_read(fname)
aoi <- sf[1, ]
area_proj(aoi)
}
