% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_accessibility.R
\name{get_accessibility}
\alias{get_accessibility}
\title{Get Accessibility to cities raster datasets}
\usage{
get_accessibility(range, index, path)
}
\arguments{
\item{range}{range of urban population}

\item{index}{index in which the file is stored in the database cloud}

\item{path}{path location}
}
\value{
copernicus land cover rasters
}
\description{
Download Accessibility to cities global rasters from the collection of 11 different layers for the year from 2015
}
\examples{
f <- "../../Om/test/" # path to the folder where you want to download the rasters
get_accessibility("5k_10k", 14189840, f)
}
