% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_accessibility.R
\name{get_accessibility}
\alias{get_accessibility}
\title{Get Accessibility to cities raster datasets}
\usage{
get_accessibility(range = NULL, index = NULL, file_path = NULL)
}
\arguments{
\item{range}{range of urban population}

\item{index}{index in which the file is stored in the database cloud}

\item{file_path}{path location}
}
\value{
accessibility rasters
}
\description{
Download Accessibility to cities global rasters from the collection of 11 different layers for the year from 2015
}
\details{
The list of range of population layer and corresponding index to use:
"5k_10k", 14189840
"10k_20k", 14189837
"20k_50k", 14189831
"50k_100k", 14189825
"100k_200k", 14189819
"200k_500k", 14189816
"500k_1mio", 14189810
"1mio_5mio", 14189807
"50k_50mio", 14189849
"5k_110mio", 14189852
"20k_110mio", 14189843
}
\examples{

\donttest{
file_path <- ".." # path to the folder where you want to download the rasters
get_accessibility(range="5k_10k", index=14189840, file_path=file_path)
}

}
