% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_copernicus_land_cover.R
\name{get_copernicus_land_cover}
\alias{get_copernicus_land_cover}
\title{Get Copernicus Global Land Cover}
\usage{
get_copernicus_land_cover(y, g, f)
}
\arguments{
\item{y}{year (2015 to 2019)}

\item{g}{Latlon grid}

\item{f}{path}
}
\value{
copernicus land cover rasters
}
\description{
Download Copernicus Global Land Cover rasters with 23 discrete classifications for the desired year from 2015 to 2019 and for the desired grid
}
\examples{
f <- "../../" # path to the folder where you want to download the rasters
get_copernicus_land_cover(2015, "W120N40", f)
}
