% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gadm_gpkg.R
\name{get_gadm_gpkg}
\alias{get_gadm_gpkg}
\title{Get GADM country level datasets (geopackage)}
\usage{
get_gadm_gpkg(x, f)
}
\arguments{
\item{x}{ISO3 country code}

\item{f}{filepath}
}
\value{
gadm geopackage
}
\description{
Download country level administrative datasets in geopackage
}
\examples{
# give path to your destination folder
f <- "../../"
# call the function
get_gadm_gpkg("NPL", f)

}
