% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gadm.R
\name{get_gadm_shp}
\alias{get_gadm_shp}
\title{Get GADM country level datasets (shapefile)}
\usage{
get_gadm_shp(iso = NULL, file_path = NULL)
}
\arguments{
\item{iso}{ISO3 country code}

\item{file_path}{filepath}
}
\value{
gadm shapefiles
}
\description{
Download country level administrative datasets in ESRI shapefile format
}
\examples{

\donttest{
# give path to your destination folder
file_path <- ".."
# call the function
get_gadm_shp(iso="NPL", file_path=file_path)
}

}
